Error estimation
I am a bit ignorant in these issues. I introduced a flag called benchmark (lack of fantasy, I know) and a few useful addition.
I introduce in each dimensional equation a way to test if the computed stress is comparable with not-dimensional one. Most of the time the error is within the standard machine precision (e-15). I introduce a formula to interpolate the vectors produced by the dimensional and not dimensional one in the respective time vectors and compute an estimation of the errors.

In general computing the mean of these difference yields 1e-4 as absolute error, and in general, cannot be associated to the way in which is compute the stress or effective strain.