Code seems to be robust
The code seems to be robust. The error systematically depends on the parameter, on the other hand, it is not something related to the internal functions and it may be the solver it self that creates a bit of problem.
If i change the combination of tollerance between dimensional and adimensional solver the error changes, and its magnitude is more or less one order of magnitude less.
Here i need more sophisticated suggestion on how to tune the solver option or to introduce more time step.