Long story short
I know that creating a branch from a branch is not exactly a good practice. On the other hand, I forgot to merge to main, and I started working on the new implementation.

Problem: I think that is impossible to have a good solution here. The main issue is with viscosity of the upper mantle and how to deal with it.

1) I have a dDdt, -> i use a guess viscosity (i.e. the diffusion creep one) -> compute tau-> compute the viscosity again -> compute tau. This approach blow up the computation, as if I start with larger viscosity i will have large stresses for a given viscosity and thus it starts to producing unrealistic values. The only alternative way is to use the previous timestep viscosity, but, it appears that is impossible to carry out this information in the ODE solver, i think that rather being ODE is a PDE what i'm going to solve.