Improvement optimisation
So, I implemented a fetch parameter that has an initial guess on the effective exponent of the slab due to the drag forces.
However, it is not satisfying. I think that an other component that drives the misfit its the following: in 2D the stress at the top of the slab IS NOT the reference stress (due to the viscosity of the mantle, and due to the flexure of the lithosphere)